model/pyproject.toml
[tool.poetry]
name = "fydp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-docx = "^1.1.2"
transformers = "^4.42.3"
torch = "^2.3.1"
tqdm = "^4.66.4"
chromadb = "^0.5.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
